Use a modal dialog for tooltips
This gives us access to some tether options that make it much easier to
position the tooltip well. In order to deal with variable space and
screen sizes I pinned the dialog so that it will remain mostly in place,
but relatively close to the item the mouse it over.