Use the official in-element helper
The private {{-in-element}} has been promoted to public api with the
addition of the insertBefore option. Not specifying this option causes
the tooltip to replace the content of the application (which is not
good). A polyfill is used to provide backwards compatible support for
versions of ember that do not have this API.