Convert to v2 addon
I used the blueprint at `embroider-build/addon-blueprint` to generate a
fresh addon and then moved all of our existing code into it. There are
some significant differences in a bunch of places to the way we had this
setup, but I think that one time drift is OK to align with the
blueprint.