Remove ember-render-modifiers
This pattern isn't encouraged anymore. Instead of using the modifier to
track changes I've added a side-effect to a getter which will repaint
the chart when the passed data is changed.

This is probably a hack, but it works.

Also, as we can now cache the getter based on data the chart won't
re-paint when it is re-sized, only when the data is modified. This
removes an annoying re-start of the animation on things like the donut
chart that would re-draw when the size of the container changed.