Enhance Integration Tests
Covered a lot more of the internals for these rendered charts in the
integration tests. It's a bit verbose, but should be easier to spot
issues in the future.
Discovered that the 100% in these tests was causing the chart to mostly
render off screen which made the percy shots mostly useless.