Improve Transition Completion
It turns out that the `end` event we were tracking here from D3 doesn't
apply to the entire transition. Each section of the donut or pie is
animated individually and end gets called for each one.

I've added a counter which gets incremented as each section animation
starts and decremented when they finish. When there are none left we can
go ahead and mark loading as complete and update the class to loaded.