Replace Loading Logic
Still not getting a reliable response from our loading indicator and
causing percy to shoot too soon. Instead we can store the promise when
we create a transition that will resolve when it is complete. If that
promise isn't set yet, or isn't resolved we're in a 'loading' state.

Had the added benefit of hiding the defensive programming around a
destroyed component. When the component is destroyed this getter won't
fire anyway.